requirements.txt
# =================================================================
#   A.R.I.S.E. - Python Dependencies (Production)
#   Install: pip install -r requirements.txt
# =================================================================

# --- Core Framework ---
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# --- Security ---
bcrypt==5.0.0
PyJWT==2.10.1
flask-cors==6.0.2
Flask-Limiter==4.1.1
limits==5.8.0

# --- Database & Data ---
openpyxl==3.1.5
et_xmlfile==2.0.0

# --- Scheduling ---
APScheduler==3.11.0
tzdata==2025.2
tzlocal==5.3.1

# --- Analytics & Visualization ---
matplotlib==3.10.8
numpy==2.4.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
pillow==12.1.0
pyparsing==3.3.2

# --- Configuration ---
python-dotenv==1.2.1

# --- Utilities ---
colorama==0.4.6
Deprecated==1.3.1
ordered-set==4.1.0
packaging==26.0
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.15.0
wrapt==2.1.1
